PYIC-3894 Refactor code to move business logic to separate service.
